requirements.txt
Django
djangorestframework
djangorestframework_simplejwt
Pillow